% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_Method.R
\name{keyfeat}
\alias{keyfeat}
\title{Key feature}
\usage{
keyfeat(cdata, criter)
}
\arguments{
\item{cdata}{Overlap clustering result.}

\item{criter}{A numeric indicating tetrachoric or polychoric correlation coefficient.}
}
\description{
Determine key features of each overlap clusters.
}
