% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_Method.R
\name{OverlapClust}
\alias{OverlapClust}
\title{Overlap Clustering}
\usage{
OverlapClust(data.simi)
}
\arguments{
\item{data.simi}{The similarity matrix of Bossa scores.}
}
\description{
Do overlap clustering with Bossa similarity in different levels of \code{p}.
}
