% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_Method.R
\name{testlmc}
\alias{testlmc}
\title{Merge test}
\usage{
testlmc(data, r)
}
\arguments{
\item{data}{A data.frame with two columns with dichotomous or polytomous.}

\item{r}{A numeric indicating tetrachoric or polychoric correlation coefficient.}
}
\description{
Do a test when try to merge two overlap clusters.
}
