% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bossa_Simi.R
\name{BossaClust}
\alias{BossaClust}
\title{Bind two factors}
\usage{
BossaClust(data, alpha = 1, p = c(0.9, 0.8, 0.7, 0.5), lin = 0.2,
  is.pca = TRUE, pca.sum.prop = 0.95, fix.pca.comp = FALSE, n.comp = 50,
  cri = 1, lintype = "ward.D2")
}
\arguments{
\item{a}{factor}

\item{b}{factor}
}
\value{
factor
}
\description{
Create a new factor from two existing factors, where the new factor's levels
are the union of the levels of the input factors.
}
\examples{
#' fbind(iris$Species[c(1, 51, 101)], PlantGrowth$group[c(1, 11, 21)])
}
